+++ src/ui/dialog/selectordialog.h
@@ -15,13 +15,14 @@
 #ifndef SELECTORDIALOG_H
 #define SELECTORDIALOG_H
 
-#include <ui/widget/panel.h>
-#include <gtkmm/treeview.h>
-#include <gtkmm/treestore.h>
-#include <gtkmm/scrolledwindow.h>
 #include <gtkmm/dialog.h>
-#include <gtkmm/treeselection.h>
 #include <gtkmm/paned.h>
+#include <gtkmm/scrolledwindow.h>
+#include <gtkmm/treemodelfilter.h>
+#include <gtkmm/treeselection.h>
+#include <gtkmm/treestore.h>
+#include <gtkmm/treeview.h>
+#include <ui/widget/panel.h>
 
 #include "ui/dialog/desktop-tracker.h"
 
@@ -52,7 +53,6 @@ public:
     SelectorDialog operator=(SelectorDialog const &d) = delete;
 
     static SelectorDialog &getInstance() { return *new SelectorDialog(); }
-
   private:
     // Monitor <style> element for changes.
     class NodeObserver;
@@ -74,12 +74,14 @@ public:
             add(_colType);
             add(_colObj);
             add(_colProperties);
+            add(_colVisible);
         }
         Gtk::TreeModelColumn<Glib::ustring> _colSelector;       // Selector or matching object id.
         Gtk::TreeModelColumn<bool> _colExpand;                  // Open/Close store row.
         Gtk::TreeModelColumn<gint> _colType;                    // Selector row or child object row.
         Gtk::TreeModelColumn<std::vector<SPObject *> > _colObj; // List of matching objects.
         Gtk::TreeModelColumn<Glib::ustring> _colProperties;     // List of properties.
+        Gtk::TreeModelColumn<bool> _colVisible;                 // Make visible or not.
     };
     ModelColumns _mColumns;
 
@@ -98,16 +100,16 @@ public:
         void on_row_deleted(const TreeModel::Path& path) override;
 
     public:
-        static Glib::RefPtr<SelectorDialog::TreeStore> create(SelectorDialog *selectordialog);
+      static Glib::RefPtr<SelectorDialog::TreeStore> create(SelectorDialog *styledialog);
 
     private:
         SelectorDialog *_selectordialog;
     };
 
     // TreeView
-    Gtk::TreeView _treeView;
+    Glib::RefPtr<Gtk::TreeModelFilter> _modelfilter;
     Glib::RefPtr<TreeStore> _store;
-
+    Gtk::TreeView _treeView;
     // Widgets
     Gtk::Paned _paned;
     Gtk::Box   _mainBox;